dockerized the application
