+++ apis/Dockerfile
@@ -1,23 +1,23 @@
 # Use an official Python runtime as a parent image
 FROM python:3.8-slim
 
-# Set the working directory to /app
+# Set the working directory in the container
 WORKDIR /app
 
 # Copy the current directory contents into the container at /app
 COPY . /app
 
-# Install any needed packages specified in requirements.txt
-RUN pip3 install --no-cache-dir -r requirements.txt
+# Copy .env file to the working directory
+COPY .env /app
 
-# Copy .env file to the container
-COPY .env .env
+# Install any needed dependencies specified in requirements.txt
+RUN pip install --no-cache-dir -r requirements.txt
 
-# Make port 80 available to the world outside this container
-EXPOSE 80
+# Make port 5000 available to the world outside this container
+EXPOSE 5000
 
 # Define environment variable
-# ENV NAME World
+ENV FLASK_APP=endpoints.py
 
 # Run app.py when the container launches
-CMD ["python", "endpoints.py"]
+CMD ["flask", "run", "--host=0.0.0.0"]
